glad: implement much more compact code generation
Instead of doing what is effectively a fully unrolled loop to load the
function pointers, index the functions and fill them by index in a loop.
This results in a much smaller code size (a couple hundred KiB with
every extension and API enabled).

Also implement early support for MX on WGL/GLX.

Signed-off-by: Steven Noonan <steven@uplinklabs.net>